server/requirements.txt
bidict==0.22.1
click==8.1.7
colorzero==2.0
Flask==2.1.3
Flask-Cors==4.0.0
Flask-JWT==0.3.2
Flask-SocketIO==5.3.6
gevent==23.9.1
gevent-websocket==0.10.1
gpiozero==1.6.2
greenlet==2.0.2
h11==0.14.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
PyJWT==1.7.1
python-dotenv==1.0.0
python-engineio==4.7.1
python-socketio==5.9.0
simple-websocket==0.10.1
wakeonlan==3.0.0
Werkzeug==2.0.3
wsproto==1.2.0
zope.event==5.0
zope.interface==6.0
